# Conflicts:
#	cmake/Modules/CocosConfigDefine.cmake
#	cmake/Modules/CocosConfigDepend.cmake
#	cmake/Modules/PreventInSourceBuilds.cmake
#	cmake/README.md
#	cocos/2d/CCLayer.h
#	cocos/2d/CCRenderTexture.cpp
#	cocos/3d/CC3DProgramInfo.cpp
#	cocos/3d/CC3DProgramInfo.h
#	cocos/3d/CCBundle3DData.cpp
#	cocos/3d/CCVertexAttribBinding.cpp
#	cocos/base/CCEnumClass.h
#	cocos/base/ccConstants.h
#	cocos/editor-support/cocostudio/ActionTimeline/CCBoneNode.h
#	cocos/physics3d/CCPhysics3DDebugDrawer.cpp
#	cocos/platform/mac/CCGLViewImpl-mac.h
#	cocos/platform/mac/CCGLViewImpl-mac.mm
#	cocos/renderer/CCTexture2D.cpp
#	cocos/renderer/CCTexture2D.h
#	cocos/renderer/CCTextureUtils.cpp
#	cocos/renderer/CCTextureUtils.h
#	cocos/renderer/backend/Buffer.h
#	cocos/renderer/backend/CommandBuffer.cpp
#	cocos/renderer/backend/CommandBuffer.h
#	cocos/renderer/backend/DepthStencilState.cpp
#	cocos/renderer/backend/DepthStencilState.h
#	cocos/renderer/backend/Device.h
#	cocos/renderer/backend/DeviceInfo.h
#	cocos/renderer/backend/Program.h
#	cocos/renderer/backend/ProgramCache.cpp
#	cocos/renderer/backend/ProgramCache.h
#	cocos/renderer/backend/ProgramState.cpp
#	cocos/renderer/backend/ProgramState.h
#	cocos/renderer/backend/RenderPassDescriptor.cpp
#	cocos/renderer/backend/ShaderCache.cpp
#	cocos/renderer/backend/ShaderCache.h
#	cocos/renderer/backend/ShaderModule.h
#	cocos/renderer/backend/Texture.cpp
#	cocos/renderer/backend/Texture.h
#	cocos/renderer/backend/Types.h
#	cocos/renderer/backend/VertexLayout.h
#	cocos/renderer/backend/metal/BufferMTL.h
#	cocos/renderer/backend/metal/BufferMTL.mm
#	cocos/renderer/backend/metal/CommandBufferMTL.h
#	cocos/renderer/backend/metal/CommandBufferMTL.mm
#	cocos/renderer/backend/metal/DepthStencilStateMTL.h
#	cocos/renderer/backend/metal/DepthStencilStateMTL.mm
#	cocos/renderer/backend/metal/DeviceInfoMTL.h
#	cocos/renderer/backend/metal/DeviceInfoMTL.mm
#	cocos/renderer/backend/metal/DeviceMTL.h
#	cocos/renderer/backend/metal/DeviceMTL.mm
#	cocos/renderer/backend/metal/ProgramMTL.h
#	cocos/renderer/backend/metal/ProgramMTL.mm
#	cocos/renderer/backend/metal/RenderPipelineMTL.h
#	cocos/renderer/backend/metal/RenderPipelineMTL.mm
#	cocos/renderer/backend/metal/ShaderModuleMTL.h
#	cocos/renderer/backend/metal/ShaderModuleMTL.mm
#	cocos/renderer/backend/metal/TextureMTL.h
#	cocos/renderer/backend/metal/TextureMTL.mm
#	cocos/renderer/backend/metal/Utils.h
#	cocos/renderer/backend/metal/Utils.mm
#	cocos/renderer/backend/opengl/BufferGL.cpp
#	cocos/renderer/backend/opengl/BufferGL.h
#	cocos/renderer/backend/opengl/CommandBufferGL.cpp
#	cocos/renderer/backend/opengl/CommandBufferGL.h
#	cocos/renderer/backend/opengl/DepthStencilStateGL.cpp
#	cocos/renderer/backend/opengl/DeviceGL.cpp
#	cocos/renderer/backend/opengl/DeviceGL.h
#	cocos/renderer/backend/opengl/DeviceInfoGL.cpp
#	cocos/renderer/backend/opengl/DeviceInfoGL.h
#	cocos/renderer/backend/opengl/ProgramGL.cpp
#	cocos/renderer/backend/opengl/ProgramGL.h
#	cocos/renderer/backend/opengl/RenderPipelineGL.cpp
#	cocos/renderer/backend/opengl/ShaderModuleGL.cpp
#	cocos/renderer/backend/opengl/TextureGL.cpp
#	cocos/renderer/backend/opengl/TextureGL.h
#	cocos/renderer/backend/opengl/UtilsGL.cpp
#	cocos/renderer/backend/opengl/UtilsGL.h
#	cocos/renderer/ccShader_Label.vert
#	cocos/renderer/ccShader_PositionTextureColor.frag
#	cocos/renderer/ccShader_PositionTextureColor.vert
#	cocos/renderer/ccShader_Position_uColor.frag
#	cocos/renderer/shaders/3D_colorTexture.frag
#	cocos/renderer/shaders/3D_terrain.frag
#	cocos/renderer/shaders/3D_terrain.vert
#	cocos/renderer/shaders/label_distanceNormal.frag
#	cocos/renderer/shaders/label_outline.frag
#	cocos/renderer/shaders/lineColor.frag
#	cocos/renderer/shaders/lineColor.vert
#	cocos/renderer/shaders/positionColor.frag
#	cocos/renderer/shaders/positionTextureColor.vert
#	cocos/renderer/shaders/ui_Gray.frag
#	cocos/scripting/deprecated/CMakeLists.txt
#	cocos/scripting/lua-bindings/auto/lua_cocos2dx_auto.cpp
#	cocos/scripting/lua-bindings/auto/lua_cocos2dx_backend_auto.cpp
#	cocos/scripting/lua-bindings/auto/lua_cocos2dx_backend_auto.hpp
#	cocos/scripting/lua-bindings/manual/LuaBasicConversions.cpp
#	cocos/scripting/lua-bindings/manual/cocos2d/lua_cocos2dx_manual.cpp
#	docs/framework_architecture_v4.png
#	external/config.json
#	templates/lua-template-default/frameworks/runtime-src/proj.ios_mac/mac/main.cpp
#	tests/.gitignore
#	tests/cpp-empty-test/Classes/HelloWorldScene.cpp
#	tests/cpp-empty-test/Resources/ipad/grossini.png
#	tests/cpp-empty-test/Resources/ipad/uv_checker_01.jpg
#	tests/cpp-tests/Classes/Camera3DTest/Camera3DTest.cpp
#	tests/cpp-tests/Classes/ClippingNodeTest/ClippingNodeTest.cpp
#	tests/cpp-tests/Classes/RenderTextureTest/RenderTextureTest.cpp
#	tests/cpp-tests/Classes/ShaderTest/ShaderTest.cpp
#	tests/cpp-tests/Classes/Sprite3DTest/Sprite3DTest.cpp
#	tests/cpp-tests/Classes/SpriteTest/SpriteTest.cpp
#	tests/cpp-tests/Classes/UITest/CocoStudioGUITest/UIVideoPlayerTest/UIVideoPlayerTest.cpp
#	tests/cpp-tests/Resources/Particles/ButterFly.plist
#	tests/cpp-tests/Resources/Particles/ButterFlyYFlipped.plist
#	tests/cpp-tests/Resources/Particles/SpinningPeas.plist
#	tests/cpp-tests/Resources/Particles/SpookyPeas.plist
#	tests/cpp-tests/Resources/Particles/particle.png
#	tests/lua-tests/project/CMakeLists.txt
#	tests/lua-tests/src/Camera3DTest/Camera3DTest.lua
#	tests/lua-tests/src/DownloaderTest/DownloaderTest.lua
#	tests/lua-tests/src/NodeTest/NodeTest.lua
#	tools/bindings-generator
#	tools/cocos2d-console

+++ tests/cpp-tests/Classes/UITest/CocoStudioGUITest/UIEditBoxTest.cpp
@@ -67,7 +67,7 @@ bool UIEditBoxTest::init()
         _editName->setPlaceHolder("Name:");
         _editName->setPlaceholderFontColor(Color3B::WHITE);
         _editName->setMaxLength(8);
-        _editName->setFontSize(editBoxSize.height/2);
+        _editName->setFontSize((int)editBoxSize.height/2);
         _editName->setText("v👐👊💝");
         _editName->setReturnType(ui::EditBox::KeyboardReturnType::DONE);
         _editName->setDelegate(this);
@@ -78,7 +78,7 @@ bool UIEditBoxTest::init()
                                         "cocosui/animationbuttonpressed.png");
         auto buttonSize = button->getContentSize();
         button->setTitleText("Single Line");
-        button->setPosition(_editName->getPosition() + Vec2(editBoxSize.width/2 + buttonSize.width/2, 0 ));
+        button->setPosition(_editName->getPosition() + Vec2(editBoxSize.width/2 + buttonSize.width/2, 0.0f));
         button->addClickEventListener([=](Ref* ref){
             _editName->setInputMode(ui::EditBox::InputMode::SINGLE_LINE);
         });
@@ -93,14 +93,14 @@ bool UIEditBoxTest::init()
         _editPassword->setMaxLength(6);
         _editPassword->setInputFlag(ui::EditBox::InputFlag::PASSWORD);
         _editPassword->setInputMode(ui::EditBox::InputMode::SINGLE_LINE);
-        _editPassword->setFontSize(editBoxSize.height/2);
+        _editPassword->setFontSize((int)editBoxSize.height/2);
         _editPassword->setDelegate(this);
         _editPassword->setVisible(true);
         addChild(_editPassword);
        
         auto buttonPassword = (ui::Button*)button->clone();
         buttonPassword->setTitleText("Multiline");
-        buttonPassword->setPosition(_editPassword->getPosition() + Vec2(editBoxSize.width/2 + buttonSize.width/2, 0 ));
+        buttonPassword->setPosition(_editPassword->getPosition() + Vec2(editBoxSize.width/2 + buttonSize.width/2, 0.0f));
         buttonPassword->addClickEventListener([=](Ref* ref){
             _editPassword->setInputMode(ui::EditBox::InputMode::ANY);
         });
@@ -126,7 +126,7 @@ bool UIEditBoxTest::init()
         
         auto buttonEmail = (ui::Button*)button->clone();
         buttonEmail->setTitleText("Multiline");
-        buttonEmail->setPosition(_editEmailParent->getPosition() + Vec2(editBoxSize.width/2 + buttonSize.width/2, 0 ));
+        buttonEmail->setPosition(_editEmailParent->getPosition() + Vec2(editBoxSize.width/2 + buttonSize.width/2, 0.0f));
         buttonEmail->addClickEventListener([=](Ref* ref){
             _editEmail->setInputMode(ui::EditBox::InputMode::ANY);
         });
@@ -190,7 +190,7 @@ bool UIEditBoxTestToggleVisibility::init()
         addChild(_TTFShowEditReturn);
         
         
-        auto editBoxSize = Size(visibleSize.width - 100, visibleSize.height * 0.1);
+        auto editBoxSize = Size(visibleSize.width - 100, visibleSize.height * 0.1f);
         
         // top
         std::string pNormalSprite = "extensions/green_edit.png";
@@ -200,7 +200,7 @@ bool UIEditBoxTestToggleVisibility::init()
         _editName->setPlaceHolder("Name:");
         _editName->setPlaceholderFontColor(Color3B::WHITE);
         _editName->setMaxLength(8);
-        _editName->setFontSize(editBoxSize.height/2);
+        _editName->setFontSize((int)editBoxSize.height/2);
         _editName->setText("v👐👊💝");
         _editName->setReturnType(ui::EditBox::KeyboardReturnType::DONE);
         _editName->setDelegate(this);
@@ -211,7 +211,7 @@ bool UIEditBoxTestToggleVisibility::init()
                                         "cocosui/animationbuttonpressed.png");
         auto buttonSize = button->getContentSize();
         button->setTitleText("Toggle Visibility");
-        button->setPosition(_editName->getPosition() + Vec2(editBoxSize.width/2 + buttonSize.width/2, 0 ));
+        button->setPosition(_editName->getPosition() + Vec2(editBoxSize.width/2 + buttonSize.width/2, 0.0f));
         button->addClickEventListener([=](Ref* ref){
             _editName->setVisible(!_editName->isVisible());
         });
@@ -225,14 +225,14 @@ bool UIEditBoxTestToggleVisibility::init()
         _editPassword->setMaxLength(6);
         _editPassword->setInputFlag(ui::EditBox::InputFlag::PASSWORD);
         _editPassword->setInputMode(ui::EditBox::InputMode::SINGLE_LINE);
-        _editPassword->setFontSize(editBoxSize.height/2);
+        _editPassword->setFontSize((float)editBoxSize.height/2);
         _editPassword->setDelegate(this);
         _editPassword->setVisible(true);
         addChild(_editPassword);
         
         auto buttonPassword = (ui::Button*)button->clone();
         buttonPassword->setTitleText("Toggle Visibility");
-        buttonPassword->setPosition(_editPassword->getPosition() + Vec2(editBoxSize.width/2 + buttonSize.width/2, 0 ));
+        buttonPassword->setPosition(_editPassword->getPosition() + Vec2(editBoxSize.width/2 + buttonSize.width/2, 0.0f));
         buttonPassword->addClickEventListener([=](Ref* ref){
             _editPassword->setVisible(!_editPassword->isVisible());
         });
@@ -258,7 +258,7 @@ bool UIEditBoxTestToggleVisibility::init()
         
         auto buttonEmail = (ui::Button*)button->clone();
         buttonEmail->setTitleText("Toggle Visibility");
-        buttonEmail->setPosition(_editEmailParent->getPosition() + Vec2(editBoxSize.width/2 + buttonSize.width/2, 0 ));
+        buttonEmail->setPosition(_editEmailParent->getPosition() + Vec2(editBoxSize.width/2 + buttonSize.width/2, 0.0f));
         buttonEmail->addClickEventListener([=](Ref* ref){
             _editEmail->setVisible(!_editEmail->isVisible());
         });
@@ -307,27 +307,27 @@ bool UIEditBoxTestTextHorizontalAlignment::init() {
     const auto glview = Director::getInstance()->getOpenGLView();
     const auto visibleOrigin = glview->getVisibleOrigin();
     const auto visibleSize = glview->getVisibleSize();
-    const auto editBoxSize = Size(visibleSize.width - 100, visibleSize.height * 0.1);
+    const auto editBoxSize = Size(visibleSize.width - 100, visibleSize.height * 0.1f);
   
     const auto createEditBox = [this, editBoxSize, visibleOrigin, visibleSize](const std::string& text,
                                       const TextHAlignment alignment,
                                       const int position_y) {
         ui::EditBox* editbox = ui::EditBox::create(editBoxSize + Size(0,40), ui::Scale9Sprite::create("extensions/green_edit.png"));
-        editbox->setPosition(Vec2(visibleOrigin.x+visibleSize.width/2-50, position_y));
+        editbox->setPosition(Vec2(visibleOrigin.x+visibleSize.width/2-50, (float)position_y));
         editbox->setFontColor(Color3B::RED);
         editbox->setPlaceHolder(text.c_str());
         editbox->setPlaceholderFontColor(Color3B::WHITE);
-        editbox->setPlaceholderFontSize(editBoxSize.height/2);
-        editbox->setFontSize(editBoxSize.height/2);
+        editbox->setPlaceholderFontSize((int)editBoxSize.height/2);
+        editbox->setFontSize((int)editBoxSize.height/2);
         editbox->setText(text.c_str());
         editbox->setTextHorizontalAlignment(alignment);
         editbox->setReturnType(ui::EditBox::KeyboardReturnType::DONE);
         addChild(editbox);
     };
   
-    createEditBox("horizontal left text", TextHAlignment::LEFT, visibleOrigin.y+visibleSize.height*3/4);
-    createEditBox("horizontal center text", TextHAlignment::CENTER, visibleOrigin.y+visibleSize.height/2);
-    createEditBox("horizontal right text", TextHAlignment::RIGHT, visibleOrigin.y+visibleSize.height/4);
+    createEditBox("horizontal left text", TextHAlignment::LEFT, (int)(visibleOrigin.y+visibleSize.height*3/4));
+    createEditBox("horizontal center text", TextHAlignment::CENTER, (int)(visibleOrigin.y+visibleSize.height/2));
+    createEditBox("horizontal right text", TextHAlignment::RIGHT, (int)(visibleOrigin.y+visibleSize.height/4));
   
     return true;
 }
@@ -341,15 +341,15 @@ bool UIEditBoxTestPressedAndDisabled::init() {
     auto glview = Director::getInstance()->getOpenGLView();
     auto visibleOrigin = glview->getVisibleOrigin();
     auto visibleSize = glview->getVisibleSize();
-    const auto editBoxSize = Size(visibleSize.width - 100, visibleSize.height * 0.1);
+    const auto editBoxSize = Size(visibleSize.width - 100, visibleSize.height * 0.1f);
 
     ui::EditBox* editbox = ui::EditBox::create(editBoxSize + Size(0,40), "extensions/yellow_edit.png", "extensions/orange_edit.png");
     editbox->setPosition(Vec2(visibleOrigin.x+visibleSize.width/2, visibleOrigin.y+visibleSize.height*3/4));
     editbox->setFontColor(Color3B::GREEN);
     editbox->setPlaceHolder("Test Pressed Image");
     editbox->setPlaceholderFontColor(Color3B::WHITE);
-    editbox->setPlaceholderFontSize(editBoxSize.height/2);
-    editbox->setFontSize(editBoxSize.height/2);
+    editbox->setPlaceholderFontSize((int)editBoxSize.height/2);
+    editbox->setFontSize((int)editBoxSize.height/2);
     editbox->setReturnType(ui::EditBox::KeyboardReturnType::DONE);
     addChild(editbox);
 
@@ -358,8 +358,8 @@ bool UIEditBoxTestPressedAndDisabled::init() {
     editbox->setFontColor(Color3B::GREEN);
     editbox->setPlaceHolder("Test Disabled Image - green");
     editbox->setPlaceholderFontColor(Color3B::WHITE);
-    editbox->setPlaceholderFontSize(editBoxSize.height/2);
-    editbox->setFontSize(editBoxSize.height/2);
+    editbox->setPlaceholderFontSize((int)editBoxSize.height/2);
+    editbox->setFontSize((int)editBoxSize.height/2);
     editbox->setReturnType(ui::EditBox::KeyboardReturnType::DONE);
     addChild(editbox);
     editbox->setEnabled(false);
@@ -369,8 +369,8 @@ bool UIEditBoxTestPressedAndDisabled::init() {
     editbox->setFontColor(Color3B::GREEN);
     editbox->setPlaceHolder("Disabled Gray effect");
     editbox->setPlaceholderFontColor(Color3B::WHITE);
-    editbox->setPlaceholderFontSize(editBoxSize.height/2);
-    editbox->setFontSize(editBoxSize.height/2);
+    editbox->setPlaceholderFontSize((int)editBoxSize.height/2);
+    editbox->setFontSize((int)editBoxSize.height/2);
     editbox->setReturnType(ui::EditBox::KeyboardReturnType::DONE);
     addChild(editbox);
     editbox->setEnabled(false);
